Adding support for pylint to the project and fixing up further lint errors.
Added pylint to the PREUPLOAD.cfg and a customized project local pylintrc
file to work around the need to reformat the whole avbtool to standard 4
spaces for indentation until all the AFTL changes are in. Also added a symlink
avbtool.py which links to the original avbtool in order for pylint to actually
detect and scan the avbtool during upload.

No changes to the functionality of avbtool.

Test: out/host/linux-x86/nativetest/libavb_host_unittest/libavb_host_unittest
Change-Id: I177e971eba450a68204f6009b4fa498f5de2cef8