Rename AftlDescriptor to AftlImage
The current AftlDescriptor is not a standard libavb descriptor as
described by AvbDescriptorTag. Rename the structure to avoid further
confusion. Rename AftlIcpHeader to AftlImageHeader, as this header
is for the whole image and not just one ICP. Update parse_aftl_image and
free_aftl_image to use the parameter "image" instead "aftl_image" to
limit the verbosity.

Bug: 153137210
Test: atest --host libavb_host_unittest
Change-Id: Ie62d31affba0b398b10a30c0abc78afdf328760f