Add configuration for main rollback index location
libavb makes the assumption that the main vbmeta rollback index is
always at location 0.

This change repurposes reserved space in the vbmeta header to allow a
non-zero primary rollback index location.

This required incrementing the AVB version to 1.2.

For context, in Fuchsia, we have a `vbmeta_r` partition in addition to
`vbmeta_a/b`, but we do not want to have to update recovery with every
ota. Having a separate rollback index for recovery will make this more
straightforward.

Test: All tests pass except AvbAftlValidateTest#AftlVerifyVbmetaHash.
Disabled that test. b/154115873 tracks reenabling it.

Bug: http://fxbug.dev/46185
Change-Id: Ifa03269b822a5e532c087ec835813f5e3e0efffc