Fixing aftltool_test breakage that was introduced by aosp/1284715.
Changed the unit test to consume the version information directly from
avbtool instead of hard-coding it. This should avoid future updates to
the tests when AVB versions are bumped up.

We are working as part of b/154318308 and b/154205081 to get the test
suites back to presubmit to avoid future breakages like this.

Also did some further Python 3 related clean up. This change is still
compatible with Python 2.

Bug: http://fxbug.dev/46185
Bug: 153057071
Test: external/avb$ atest :all
Test: ./aftltool_integration_test.py
Change-Id: I79900e746e7264a1a118f55dffed01583fd8731a