Changed avbtool to be run with the embedded launcher.
This will package avbtool with the prebuilt Python 3 runtime (py3-cmd)
which ensures that avbtool will execute independently of the Python
runtime that is installed on the system where Android is built.

Note: The embedded_launcher is set by default to false and thus we have
to explicitly enable it in the Android.bp file for py_binary_host host
targets.

Bug: 155100965
Bug: 151336743
Test: external/avb$ atest :all
Change-Id: I6a61c8cb7f3f5aebc962843376a98b119eae54f5