Unpack AftlIcpEntry structure
Having an unpacked structure allows us to dereference pointer members
without running the risk of violating any alignment rule.  We
previously did a partial read on the header to determine the size of the
proofs, declared them as a flexible array member and packed the
structure (see b09b12a). Enforce a manual allocation of the proof member
to avoid having this extra pre-read.

Move the AftlImageHeader closer to the AftlImage definition.

Test: m libavb_host_unittest && ./out/host/linux-x86/nativetest/libavb_host_unittest/libavb_host_unittest
Test: m libavb_host_unittest && ./out/host/linux-x86/nativetest64/libavb_host_unittest/libavb_host_unittest
Bug: 154303414
Change-Id: Iccb473a4600327811e35857f884001b37afdf361