Define and use read_u{8,16,64}
A large portion of the utility functions is parsing the data structures
on disk. The pattern always involve: reading a certain size of data,
ensuring that there is enough space available; converting the value from
big-endian to host byte order and advancing the pointer pass that value.
Define new functions to capture this task. Use these functions for
parsing the TrillanLogRoot descriptor.

Test: m libavb_host_unittest && ./out/host/linux-x86/nativetest64/libavb_host_unittest/libavb_host_unittest
Test: m libavb_host_unittest && ./out/host/linux-x86/nativetest/libavb_host_unittest/libavb_host_unittest
Bug: 154303414
Change-Id: Ie23639f69804dc6a23c66a5e96c45d8c0ca7586c