Keep copy of raw TrillianLogRoot
The leaf validation requires a complete copy of the TLS-encoded version
of the TrillianLogRoot descriptor. There is a complete copy in the
AftlImage structure on disk. Previously, we only kept the parsed version
of the descriptor in memory, reencoding the whole descriptor when the raw
version was required. Drop the reencoding step by keeping the raw copy
in memory.

Bug: 152958312
Test: m libavb_host_unittest && ./out/host/linux-x86/nativetest/libavb_host_unittest/libavb_host_unittest
Test: m libavb_host_unittest && ./out/host/linux-x86/nativetest64/libavb_host_unittest/libavb_host_unittest
Change-Id: I20ffd7d8f4bc06719f9b28601f1d0e611e4f524c