libavb: Abstract crypto operations
Abstract the implementation details of crypto operations to allow them
to be swapped out at build time. The original SHA implementation is
still provided, along with a boringssl implementation for user space and
bootloaders where boringssl is available.

While we're at it, add a TEST_MAPPING file to run the unit test with
both crypto implementations in presubmit.

Test: atest --host --host-unit-test-only --test-mapping external/avb
Bug: 185329132
Change-Id: Ib391457684a1819a1bf921695ddb880da0f3967d