util: Deprecate avb_{debug,error,fatal}v() macros
Noting that both the avb_*() and avb_*v() macros use the underlying
avb_printv() function, which requires a terminating NULL sentinel,
unify the 2 classes of macros to wrap the call and encapsulate the NULL
argument; as the resulting macros are variadic, they now cover the
single-parameter case, deprecating the need for having distinct classes
of macros.

As a result, keep the avb_*() macros and deprecate the avb_*v() macros,
as they have a more convenient name, but keep the old macros defined, in
case client code uses them. Note that this new implementation of the
macros is still compatible with the (previously mandatory) use-case
where a NULL sentinel needed to be passed.

Test: -