sysdeps: Add support for vfprintf() logging
Introduce a build flag to make logging calls (avb_{debug,error,fatal})
use vfprintf(3) instead of printf(3), which greatly improves the
probability of the call resulting in a single write(2), as opposed to
our current implementation of avb_printv(), which calls fprintf(3) once
per received argument.

This is useful when stderr points to a file that is shared and/or format
entries per write(2) call, such as /dev/kmsg. Note that client code
passing the now deprecated NULL sentinel to the logging macros
avb_{debug,error,fatal}() must NOT enable this feature.

Test: -
Reported-by: JeongHyeon Lee <jhs2.lee@samsung.com>