avbtool:support append vbmeta image directly at the end of a partition
before:The original avbtool only supports appending the vbmeta data
to the end of a given image when adding a vbmeta image, and then
appends an avbfooter at the very end of the partition. This results
in the position of the vbmeta data being unfixed—it changes with
variations in the compilation process.

new change:A new methon is added here: vbmeta struct can be append
at the end of partiton. This will facilitate easier reading of the
vbmeta data, eliminating the need to parse the avbfooter.
Also a block size input is added for append_vbmeta_image

original append method:
avbtool.py append_vbmeta_image \
--image pathtoimage --partition_size 262144 --vbmeta_image vbmeta.img

new append methon:
avbtool.py append_vbmeta_image \
--image pathtoimage --partition_size 262144 --vbmeta_image vbmeta.img \
--block_size 16384 --append_to_end

Signed-off-by: jingfei <jingfei@xiaomi.com>